refactor(geo): extract shared flight-popup JS for the 3D map (#268)
