+++ src/dji_metadata_embedder/geo/flightmap_html.py
@@ -16,6 +16,7 @@ from html import escape
 from pathlib import Path
 
 from .flightmap import flights_to_geojson
+from .flightmap_js import FLIGHT_POPUP_JS
 from .tiles import DEFAULT_TILE_STYLE, tile_layer_js
 from .track import Track
 
@@ -74,39 +75,7 @@ const data = JSON.parse(document.getElementById('flight-data').textContent);
 const map = L.map('map');
 __TILE_LAYER__
 
-const esc = s => String(s).replace(/[&<>"']/g,
-  ch => ({'&':'&amp;','<':'&lt;','>':'&gt;','"':'&quot;',"'":'&#39;'}[ch]));
-
-// 12 visually distinct track colours, cycled when a folder has more flights.
-const PALETTE = ['#e6194b', '#3cb44b', '#4363d8', '#f58231', '#911eb4',
-                 '#42d4f4', '#f032e6', '#bfef45', '#469990', '#9a6324',
-                 '#800000', '#000075'];
-
-function fmtDuration(total) {
-  const h = Math.floor(total / 3600);
-  const m = Math.floor((total % 3600) / 60);
-  const s = total % 60;
-  const mm = h ? String(m).padStart(2, '0') : String(m);
-  return (h ? h + ':' + mm : mm) + ':' + String(s).padStart(2, '0');
-}
-
-function popupHtml(p) {
-  let html = `<div class="flight-popup"><b>${esc(p.name || '')}</b>`;
-  if (p.start) html += `<br>${esc(p.start)}`;
-  if (p.duration_s != null) html += `<br>duration: ${fmtDuration(p.duration_s)}`;
-  if (p.height_min != null) {
-    html += `<br>height: ${p.height_min} to ${p.height_max} m above takeoff`;
-  } else if (p.alt_min != null) {
-    html += `<br>altitude: ${p.alt_min} to ${p.alt_max} m (as logged)`;
-  }
-  html += `<br>${p.points} GPS point${p.points === 1 ? '' : 's'}`;
-  if (p.segments) {
-    html += `<br>recorded across ${p.segments.length} files: ` +
-            `${esc(p.segments[0])} → ${esc(p.segments[p.segments.length - 1])}`;
-  }
-  html += '</div>';
-  return html;
-}
+__SHARED_JS__
 
 const overlays = {};
 const allLatLngs = [];
@@ -285,7 +254,8 @@ def flights_to_html(
         css_sri=_LEAFLET_CSS_SRI,
         js_sri=_LEAFLET_JS_SRI,
         data=data,
-        app_js=_APP_JS.replace("__TILE_LAYER__", tile_layer_js(tile_style)),
+        app_js=_APP_JS.replace("__TILE_LAYER__", tile_layer_js(tile_style))
+        .replace("__SHARED_JS__", FLIGHT_POPUP_JS),
     )
